where did my error messages go...
git-svn-id: https://svn.salilab.org/imp/trunk@2987 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/domino/include/JunctionTree.h
@@ -32,16 +32,16 @@ public:
   void set_nodes(int number_of_nodes);
 
   void add_edge(int v1,int v2) {
-    IMP_assert(v1 < boost::num_vertices(g_),
+    IMP_assert(static_cast<unsigned int>(v1) < boost::num_vertices(g_),
                "input node index (" << v1 << ") is out of range ("
                << boost::num_vertices(g_) << std::endl);
-    IMP_assert(v2 < boost::num_vertices(g_),
+    IMP_assert(static_cast<unsigned int>(v2) < boost::num_vertices(g_),
                "input node index (" << v2 << ") is out of range ("
                << boost::num_vertices(g_) << std::endl);
     boost::add_edge(v1,v2,g_);
   }
   void set_node_name(int vi, const std::string &name) {
-    IMP_check(vi < boost::num_vertices(g_),
+    IMP_check(static_cast<unsigned int>(vi) < boost::num_vertices(g_),
               "input node index (" << vi << ") is out of range ("
               << boost::num_vertices(g_) << ")"<<std::endl,ValueException);
     data_[vi].push_back(name);
@@ -50,7 +50,7 @@ public:
   int get_number_of_components(int vi) const {return data_[vi].size();}
 
   const std::string get_component_name(int vi,int ci) const {
-    IMP_check(vi < boost::num_vertices(g_),
+    IMP_check(static_cast<unsigned int>(vi) < boost::num_vertices(g_),
               "input node index (" << vi << ") is out of range ("
               << boost::num_vertices(g_) <<")" <<std::endl,ValueException);
     IMP_check(ci < get_number_of_components(vi),
